jwk: Move Abstract Base Clases into separate file.
